+++ src/components/ShopHighlight.js
@@ -3,23 +3,48 @@ import { globalContext } from "../globalContext";
 
 function ShopHighlight({ bonus }) {
     /* const upgradeActive = userData.points.net >=  */
-    const { userData } = useContext(globalContext);
-    const userBonusData = userData.bonusData[bonus.id - 1];
-    const upgradePrice = Math.ceil(bonus.price.base * (bonus.price.growthRate ** userBonusData.level));
+    const { progression, sequence, recursion, price, image, name, description, id } = bonus
+    const { userData, setUserData } = useContext(globalContext);
+    const userBonusData = userData.bonusData[id - 1];
+    const upgradePrice = Math.ceil(price.base * (price.growthRate ** userBonusData.level));
+
+    function handleUpgradeClick() {
+        if (userData.points.net >= upgradePrice) {
+            console.log('Upgrade allowed');
+
+            const updatedBonusArray = userData.bonusData.map((bonus,index) => {
+                return index === id - 1
+                    ? {...bonus,
+                        level: bonus.level + 1
+                    }
+                    : bonus
+            })
+
+            setUserData(data => {
+                return {...data,
+                    bonusData: updatedBonusArray,
+                    points: {...data.points,
+                        spent: data.points.spent + upgradePrice}
+                }
+            })
+        } else console.log('Insufficient funds');
+    }
+    
+
     let nextValue;
-    if (bonus.progression === "sequence") nextValue = bonus.sequence[userBonusData.level + 1];
-    if (bonus.progression === "increment") nextValue = userBonusData.level + 1;
-    if (bonus.progression === "recursive") {
-        nextValue = bonus.recursion.base * (bonus.recursion.growthRate ** userBonusData.level);
+    if (progression === "sequence") nextValue = sequence[userBonusData.level + 1];
+    if (progression === "increment") nextValue = userBonusData.level + 1;
+    if (progression === "recursive") {
+        nextValue = recursion.base * (recursion.growthRate ** userBonusData.level);
     }
 
     return (
         <div id="shop-highlight">
-            <img src={bonus.image} />
-            <h1>{bonus.name}</h1>
-            <p>{bonus.description}</p>
+            <img src={image} />
+            <h1>{name}</h1>
+            <p>{description}</p>
             <strong>{upgradePrice}</strong>
-            <button onClick={() => console.log('Hello')}>Upgrade</button>
+            <button onClick={handleUpgradeClick}>Upgrade</button>
             <em>{`(to ${nextValue})`}</em>
         </div>
     )